Massive changes:
- Get rid of VideoEngine, instead the new OutputSchedulerThread is a
thread whose main job is to synchronise the output device (viewer or
sequential writer such as FFMPEG)

- OutputSchedulerThread contains a thread-pool rendering X frames
concurrently using QRunnable

- The command-line arguments now accept frame ranges after the writer
name in the format  FirstFrame-LastFrame, e.g: 10-40, if not set then
it will be picked automatically from what the plug-in frame range is.

- After this commit there are still a few deadlocks and the background
renderer is NOT tested.

- The CPU activity tracking is not implemented, by default you can
control how many CPUs are rendering concurrently from the preferences
(nParallelRenders).

- The buffering is in the current state infinite… when doing playback
on small frame ranges it then grows very quickly, we need to reduce its
size.