Changes:
- Remove animation button on all parameters
- Rework the right click menus on all parameters and make it more
consistent according to the state of the parameter
- Rework the copy/paste functionality in the right click menu (3 modes
available: copy value, copy animation or copy links)
- Use hard-links instead of expressions when copying links or drag and
dropping parameters
- Drag and dropping parameters can now be originated from the “widget”
of the parameter (e.g the spinbox of a Double parameter) when holding
the Control key (CMD on mac).  You can drag it to another parameters
dimension of the same type. If their type differ, you have to use a
custom expression yourself to link the 2 parameters.
You can also drag and drop the label of the widget to quickly link all
dimensions of the parameter to all others of the target parameter in
which case both parameters must have the same number of dimensions.
- When drag and dropping a parameter onto another one, if CTRL and
SHIFT are held, the link will be an expression instead of the following
form:

curve(frame, dimension) * sourceParam.get()[dimension]

- Always hide Custom parameters as done on other hosts! Leaving it
displayed could cause terrible performances for some plug-ins such as
Genarts S_LensFlare